add cluster-wide unique id to children
a unique child id is needed to reliably check if a spec or child is
already present in the cluster.

To avoid checking for uniqueness when starting a child and changing the
dynamic supervisor public api (which uses pids to refer to children,
e.g. to stop it), an uuid is autogenerated and saved in the child spec
instead of the id field, which is unused